add new entities, edit exists entities

+++ src/main/java/uz/bat/model/AbstractUzBatEntity.java
@@ -15,15 +15,14 @@ public abstract class AbstractUzBatEntity implements Serializable
     private static final long serialVersionUID = 1L;
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    protected int id;
+    protected Long id;
 
-
-    public int getId()
+    public Long getId()
     {
         return id;
     }
 
-    public void setId(int id)
+    public void setId(Long id)
     {
         this.id = id;
     }
